Move substitution logic to corresponding module
